Rename code package to app
