--- confluence_openapi_tool/__init__.py
@@ -1,3 +1,0 @@
-from .server import create_app
-
-__all__ = ["create_app"]
